FIX: Address Site Traffic Explorer review feedback
The explorer split its analytics across a query object and a thin service wrapper, diverged from canonical browser detection, and retained bespoke presentation and test scaffolding.

Consolidate the report in one service so its contract, timeout handling, SQL, and response assembly share one boundary. Match BrowserDetection precedence while keeping Opera ahead of Chrome, reuse the dashboard traffic error state and shared CSS patterns, remove unneeded KPI tooltips and dead translations, and delete obsolete acceptance/debug coverage.
