+++ spec/system/admin_dashboard_redesign/site_traffic_section_spec.rb
@@ -137,8 +137,7 @@ describe "Admin Dashboard Redesign | Site Traffic section" do
     expect(traffic).to have_chart
   end
 
-  it "takes an admin to the traffic explorer with the selected period when they click See details",
-     time: Time.zone.local(2026, 5, 14, 12, 0, 0) do
+  it "takes an admin to the traffic explorer with the selected period when they click See details" do
     SiteSetting.enable_site_traffic_explorer = true
     Fabricate(:logged_in_browser_application_request, date: "2026-05-05", count: 10)
 
@@ -155,8 +154,7 @@ describe "Admin Dashboard Redesign | Site Traffic section" do
     )
   end
 
-  it "links to the aggregate traffic report by default",
-     time: Time.zone.local(2026, 5, 14, 12, 0, 0) do
+  it "links to the aggregate traffic report by default" do
     Fabricate(:logged_in_browser_application_request, date: "2026-05-05", count: 10)
 
     dashboard.visit_with_query(range: "custom", start_date: "2026-05-01", end_date: "2026-05-12")
@@ -167,8 +165,7 @@ describe "Admin Dashboard Redesign | Site Traffic section" do
     )
   end
 
-  it "keeps the traffic explorer action hidden from a moderator",
-     time: Time.zone.local(2026, 5, 14, 12, 0, 0) do
+  it "keeps the traffic explorer action hidden from a moderator" do
     sign_in(moderator)
 
     dashboard.visit_with_query(range: "custom", start_date: "2026-05-01", end_date: "2026-05-12")
